Can now create an Author entity after fixing issue with json and date. need to figure out enums

+++ src/CRUDLibrary.Data/LIB_DB/AuthorBook.cs
@@ -2,11 +2,11 @@
 {
     public partial class AuthorBook
     {
-        public decimal AuthorBookId { get; set; }
-        public decimal AuthorId { get; set; }
+        public int AuthorBookId { get; set; }
+        public int AuthorId { get; set; }
         public virtual Author? Author { get; set; }
 
-        public decimal BookId { get; set; }
+        public int BookId { get; set; }
         public virtual Book? Book { get; set; }
     }
 }
