+++ src/CRUDLibrary.Data/LIB_DB/Borrower.cs
@@ -5,7 +5,7 @@ namespace CRUDLibrary.Data.LIB_DB
 {
     public partial class Borrower
     {
-        public decimal BorrowerId { get; set; }
+        public int BorrowerId { get; set; }
         public string Name { get; set; }
 
         public virtual ICollection<BookBorrower> BookBorrows { get; set; } = new List<BookBorrower>();
