+++ src/CRUDLibrary.Domain/Models/AuthorBook/DeleteAuthorBook.cs
@@ -8,28 +8,28 @@ namespace CRUDLibrary.Domain.Models
 {
     public class DeleteAuthorBookRequest : RequestModel
     {
-        public decimal AUTHOR_BOOK_ID { get; set; }
+        public int AUTHOR_BOOK_ID { get; set; }
 
-        public decimal AUTHOR_ID { get; set; }
-        public decimal BOOK_ID { get; set; }
+        public int AUTHOR_ID { get; set; }
+        public int BOOK_ID { get; set; }
     }
 
     public class DeleteAuthorBookResponse : ResponseModel
     {
-        public decimal AUTHOR_BOOK_ID { get; set; }
+        public int AUTHOR_BOOK_ID { get; set; }
 
-        public decimal AUTHOR_ID { get; set; }
+        public int AUTHOR_ID { get; set; }
         public string AUTHOR_NAME { get; set; } = string.Empty;
         
-        public decimal BOOK_ID { get; set; }
+        public int BOOK_ID { get; set; }
         public string BOOK_TITLE { get; set; } = string.Empty;
     }
     public class DeleteAuthorBookSubmitRequest : RequestModel
     {
-        public decimal AUTHOR_BOOK_ID { get; set; }
+        public int AUTHOR_BOOK_ID { get; set; }
 
-        public decimal AUTHOR_ID { get; set; }
-        public decimal BOOK_ID { get; set; }
+        public int AUTHOR_ID { get; set; }
+        public int BOOK_ID { get; set; }
     }
 
     public class DeleteAuthorBookSubmitResponse : ResponseModel
